Data.Cruncher.FinalResult

Description

The highest level of a response that Cruncher deals with. Contains only the final result of a sandbox run, including compilation, execution, and output files (which are base64-encoded).

Documentation

data FinalResult

The final result for a given request.

This contains the Result obtained from both evaluation and compilation as well as any files which resulted from performing the above steps. Such files should placed in ~output of the evaluation.

This data type also handles error handling, in the form of types.

Constructors

FinalResult

Fields

compile :: Maybe Result

The compilation result, if any

run :: Maybe Result

The execution result, if any

outputFiles :: Map String ByteString

Base64-encoded output files.

NoSuchLanguage

SELinuxNotEnforcing