Data.Cruncher.Language

Description

Documentation

data Language

Describes what a programming language looks like internally.

Constructors

Language

Fields

codeFilename :: String

What to store the incoming file as.

compileCommand :: Maybe [Text]

What to compile with.

compileTimeout :: Maybe Int

How many seconds to give the compilation before timeout.

runCommand :: [Text]

What to run with.

runTimeout :: Int

How many seconds to give the compilation before timeout.

codemirror :: String

How does this get highlighted in CodeMirror?

rpm :: String

Which RPM provides this? Used for summary/version only.

displayName :: String

How should this language be displayed in UIs?