Data.Cruncher.Request

Description

Handles incoming requests (usually from our Yesod frontend, but could be from anywhere).

Documentation

data Request

Describes an incoming request to the system. Proper use of this will normally lead to some kind of a IO Result.

Constructors

Request

Fields

language :: String

code :: Text

files :: Maybe [(String, ByteString)]

compileOnly :: Bool

stdin :: Maybe Text