Data.Cruncher.Result

Description

Contains data types/constructors for individual sandbox runs. For example, the compile step will produce a Result, which is defined in this module. The execution/evaluation step will also produce a Result.

Documentation

data Result

Describes the result we get back after performing an evaluation (or compilation). This is almost always wrapped in IO.

Constructors

Result

Fields

stdout :: Text

stderr :: Text

wallTime :: Int

exitCode :: Int