Data.Cruncher.SELinux

Description

Provides our actual sandboxing capabilities by making heavy use of Shelly to call out to the SELinux-provided sandbox command.

Right now, it is heavily coupled to Cruncher, but it could probably be made into its own little library, without much grief, if ever necessary.

Documentation

isSELinuxEnforcing :: IO Bool

Determine whether SELinux is enforcing.

runInSandbox :: [Text] -> Int -> Maybe Text -> FilePath -> IO Result

compile :: Language -> FilePath -> IO (Maybe Result)

Compile something in a sandbox. TODO: Move hardcoded timeout to config or something.

If there is no compileCommand defined for the given Language, we return Nothing, so that this can be used in a monadic chain with no averse effects.

If there is a compileCommand, we run it and check its exitCode field, returning the appropriate side of an Either Result (Left = non-0, Right = 0).

runRequest :: Request -> IO FinalResult

Perform the entire process of compilation and execution.